% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent.R
\name{autoplot.vent}
\alias{autoplot.vent}
\title{autoplot vent}
\usage{
\method{vent}{data.frame}(dat)
}
\arguments{
\item{dat}{a vent dataframe.}
}
\value{
plot.
}
\description{
autoplot_vent plot a dataframe of class vent
}
