% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_vent.R
\name{autoplot_vent}
\alias{autoplot_vent}
\title{autoplot vent.}
\usage{
autoplot_vent(dat)
}
\arguments{
\item{dat}{a vent dataframe.}
}
\value{
plot.
}
\description{
autoplot_vent plot a dataframe of class vent
}
