% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_summary.R
\name{bind_summary}
\alias{bind_summary}
\title{bind_summary}
\usage{
bind_summary(inter = TRUE, path)
}
\arguments{
\item{inter}{logical.}

\item{path}{if inter is false, path of xlsx.}
}
\value{
an xlsx file or a dataframe
}
\description{
bind together xlsx files returned by \code{\link{summarize_vent}}
}
