% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_summary.R
\name{find_bins}
\alias{find_bins}
\title{find_bins in vent dataframes.}
\usage{
find_bins(dat, baseline, bin)
}
\arguments{
\item{dat}{vent dataframe.}

\item{baseline}{baseline duration in minutes}

\item{bin}{bin duration in minutes}
}
\value{
a vent dataframe with a column "int_min" indicating the left limit of the interval in minutes and, a
   a colum "int_sec" indicating the left and right limits of the interval in seconds. Time of injection is at time 0.
}
\description{
find_bins creates left closed duration intervals and filters a vent dataframe.
}
\note{
this function is used by s3ummarize_vent
}
