% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_session.R
\name{get_iox}
\alias{get_iox}
\title{read iox txt files.}
\usage{
get_iox(iox_data, inter = TRUE, shiny_f = FALSE)
}
\arguments{
\item{iox_data}{path to folder containing data or, list of files to import when `shin_f = TRUE`.}

\item{inter}{logical for using or not dialogs to input data and select folders.}

\item{shiny_f}{TRUE if to be used in rvent_app.}
}
\value{
a list of dataframe:
\itemize{
   \item vent: a list of dataframes of efach of the iox files in which time_s = 0 is the start of
   the session.
   \item tsd_s: dataframe of all comments with each row rappresenting a
   different subject.}
}
\description{
get_iox reads and merges files created by the software SCIREQ. It returns a dataframe. The function is used internally by
\code{\link{import_session}}.
}
\seealso{
normalizetime_vent(), import_session(), split_comments().
}
