% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iox_shiny.R
\name{get_iox_shiny}
\alias{get_iox_shiny}
\title{read iox txt files.}
\usage{
get_iox_shiny(iox_files)
}
\arguments{
\item{iox_files}{input of shiny::fileInput.}
}
\value{
a list of dataframe:
\itemize{
   \item vent: a list of dataframes of each of the iox files in which time_s = 0 is the start of
   the session.
   \item tsd_s: dataframe of all comments with each row rappresenting a
   different subject.}
}
\description{
get_iox reads and merges files created by the software SCIREQ. It returns dataframes. The function is used internally by
\code{\link{import_session}}.
}
\seealso{
normalizetime_vent(), import_session(), split_comments().
}
