% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vent.R
\name{import_iox}
\alias{import_iox}
\title{Import data ventilation.}
\usage{
import_iox(link_file)
}
\arguments{
\item{link}{Path of the file to import.}
}
\value{
A dataframe of class fightscore. Columns are:
\describe{
  \item{xx}{xxxx}
}
}
\description{
Import a txt file created by the software IOX and return a dataframe of vent.
}
\seealso{
\url{https://www.boris.unito.it/}
}
