% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vent.R
\name{import_session}
\alias{import_session}
\title{Import iox session.}
\usage{
import_session(baseline = 30)
}
\arguments{
\item{baseline}{Length of baseline in minutes.}
}
\value{
A list of dataframes of class vent.
}
\description{
Import txt files created by the software IOX and return a dataframe of class vent.
}
