% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_session.R
\name{inj_comments}
\alias{inj_comments}
\title{get the time of injection.}
\usage{
inj_comments(dat)
}
\arguments{
\item{dat}{a dataframe returned by \code{\link{get_iox}}.}
}
\value{
a dataframe tsd_s
}
\description{
inj_comments discovers the time of the injection of a dataframe returned by \code{\link{get_iox}}
by finding the position of comments (comments_tsd) containing subject, drug, dose and units.  The function is used internally
by  \code{\link{import_session}}
}
\seealso{
* [get_iox()].
* [normalizetime_vent()].
* [import_session()].
}
