% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent.R
\name{new_vent}
\alias{new_vent}
\title{vent constructor}
\usage{
new_vent(dat = data.frame())
}
\arguments{
\item{dat}{object}
}
\description{
vent constructor
}
\concept{\url{https://adv-r.hadley.nz/s3.html#s3-constructor}{advance R}}
