% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_session.R
\name{normalizetime_vent}
\alias{normalizetime_vent}
\title{normalize time of injection.}
\usage{
normalizetime_vent(dat, tsd_s, tofill, baseline)
}
\arguments{
\item{dat}{a dataframe returned by \code{\link{get_iox}}.}

\item{tsd_s}{the dataframe returned by \code{\link{inj_comments}}.}

\item{tofill}{vector of values to replace missing data in comments_tsd, used if inter = FALSE.}

\item{baseline}{Length of baseline in minutes.}
}
\value{
a dataframe.
}
\description{
normalizetime_vent tidies up a dataframe returned by \code{\link{get_iox}} and creates a column of time from injection in seconds
The function is used internally by \code{\link{import_session}}
}
\seealso{
* [get_iox()].
* [inj_comments()].
* [import_session()].
}
