% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_plots.R
\name{session_plots}
\alias{session_plots}
\title{session_plots.}
\usage{
session_plots(dat, inter = TRUE, path, baseline = 30, bin = 3)
}
\arguments{
\item{dat}{a iox dataframe.}

\item{inter}{logical that determines if dialogs will be used.}

\item{path}{where to save the files.}

\item{baseline}{length of baseline in minutes.}

\item{bin}{length of bin in minutes.}
}
\value{
plots
}
\description{
session_plots creates plots from list of vent dataframes.
}
