% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_comments.R
\name{split_comments}
\alias{split_comments}
\title{split comments.}
\usage{
split_comments(tsd, detect, sep, rem)
}
\arguments{
\item{tsd}{dataframe of comments (2 columns).}

\item{detect}{regexp for detecting rows with multiple subjects}

\item{sep}{regexp  for separating subjects from rest}

\item{rem}{the string to remove}
}
\value{
a dataframe tsd_s
}
\description{
split_comments reorganizes a dataframe with column "time" and "comments" to return a dataframe in which
each subj is in a different row. It is used internally by \code{\link{get_iox}}. Patterns are "rat 1 and 2 and 3 ..." and
"rat 1 2 3.."
}
