% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_vent.R
\name{summarize_vent}
\alias{summarize_vent}
\title{summarize vent dataframes.}
\usage{
summarize_vent(dat)
}
\arguments{
\item{dat}{a list of vent dataframes as returned by import_session.}
}
\value{
An excel file.
}
\description{
summarize_vent creates duration intervals, filters and summarize a list of iox dataframes.
}
