% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_vent_ni.R
\name{summarize_vent_ni}
\alias{summarize_vent_ni}
\title{summarize vent dataframes.}
\usage{
summarize_vent_ni(dat, basel = 30, bin = 3)
}
\arguments{
\item{dat}{a list of vent dataframes as returned by import_session.}

\item{basel}{length of baseline in minutes}

\item{bnin}{length of bins in minutes}
}
\value{
a list of dataframes including one of class vent
}
\description{
summarize_vent_in is the non-interactive version of \code{\link{summarize_vent}} and it is used internaly by plot_vent creates duration intervals, filters and summarize a list of vent dataframes.
}
