% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent.R
\name{validate_vent}
\alias{validate_vent}
\title{#' vent validator
#'
#' validate an object of class vent
#'
#' @param dat object
#' @export validate_vent}
\usage{
validate_vent(dat)
}
\description{
#' vent validator
#'
#' validate an object of class vent
#'
#' @param dat object
#' @export validate_vent
}
