% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.vent.R
\name{autoplot.vent}
\alias{autoplot.vent}
\alias{autoplot}
\title{autoplot.vent}
\usage{
autoplot.vent(dat, vent_stat = "mean", fsave = TRUE, measure = "ALL")
}
\arguments{
\item{dat}{a vent dataframe.}

\item{vent_stat}{stat used to summarize bins displayed in the y axis ("median" or "mean").}

\item{fsave}{if TRUE the function saves the figure in a pdf, otherwise returns a fig.}

\item{measure}{metric to display.}
}
\value{
plot.
}
\description{
autoplot.vent plots a vent dataframe
}
