% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent.R
\name{autoplot.vent}
\alias{autoplot.vent}
\title{#' Methods of class vent
#'
#' These functions provide methods for collection, analyzing and visualizing a
#' set of vent results from a common data set
#'
#'
#'
#' @export vent}
\usage{
\method{vent}{data.frame}(dat, fsave = TRUE)
}
\arguments{
\item{dat}{a vent dataframe.}
}
\value{
plot.
}
\description{
"vent" <- function(x, ...) UseMethod("vent")
}
\details{
#' @rdname vent
#' @method vent default
#' @export
vent.default <- function(dat) {
    if (!"data.frame" %in% class(dat)) stop("The object is not a data.frame")
}
autoplot vent

autoplot.vent plots a vent dataframe
}
