% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.vent.R
\name{autoplot.vent}
\alias{autoplot.vent}
\alias{autoplot}
\title{autoplot.vent}
\usage{
autoplot(dat = dat_vent, fsave = FALSE)
}
\arguments{
\item{dat}{a vent dataframe.}

\item{fsave}{if TRUE the function saves the figure in a pdf, otherwise returns a fig}
}
\value{
plot.
}
\description{
autoplot.vent plots a vent dataframe
}
