% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bins.R
\name{find_bins}
\alias{find_bins}
\title{find_bins in vent dataframes.}
\usage{
find_bins(dat, bin, filter_vals = TRUE)
}
\arguments{
\item{dat}{dataframe as returned by \code{\link{import_session}}.}

\item{bin}{bin duration in minutes.}

\item{filter_vals}{if TRUE filters data to eliminte impossible value for rat phisiology.}
}
\value{
a dataframe with a column "int_min" indicating the left limit of the interval in minutes and, a
   a colum "int_sec" indicating the left and right limits of the interval in seconds. Time of injection is 0.
}
\description{
find_bins creates left closed duration intervals and filters the dataframe returned by import_vent.
}
\note{
this function is used by summarize_vent
}
