% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_session.R
\name{get_iox}
\alias{get_iox}
\title{read iox txt files.}
\usage{
get_iox(iox_folder, baseline = 30, inter = TRUE)
}
\arguments{
\item{iox_folder}{path to folder for saving data, used if inter = FALSE.}

\item{baseline}{Length of baseline in minutes.}

\item{inter}{logical for using or not dialogs to input data and select folders.}
}
\value{
a dataframes.
}
\description{
get_iox reads and merges files created by the software SCIREQ. It return a dataframe. The function is used internally by
\code{\link{import_session}}
}
\seealso{
* [inj_comments()].
* [normalizetime_vent()].
* [import_session()].
}
