% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_session.R
\name{import_session}
\alias{import_session}
\title{Import iox session.}
\usage{
import_session(baseline = 30, inter = TRUE, iox_folder, comments_tsd,
  tofill)
}
\arguments{
\item{baseline}{Length of baseline in minutes.}

\item{inter}{logical for using or  not dialogs to input data and select folders.}
}
\value{
A list of dataframes of class iox.
}
\description{
Import txt files created by the software IOX and return a list of dataframes of class iox.
}
