% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_session.R
\name{normalizetime_vent}
\alias{normalizetime_vent}
\title{normalize time of injection.}
\usage{
normalizetime_vent(dat, tsd_s, tofill, baseline = 30)
}
\arguments{
\item{dat}{a dataframe returned by \code{\link{get_iox}}.}

\item{tsd_s}{the dataframe of comments returned by \code{\link{get_iox}}.}

\item{tofill}{vector of values to replace missing data in tsd_s (used if inter = FALSE).}

\item{baseline}{length of baseline in minutes.}
}
\description{
normalizetime_vent tidies up the ataframes returned by \code{\link{get_iox}} and creates a column of time from injection in seconds.
The function is used internally by \code{\link{import_session}}.
}
\seealso{
get_iox(), import_session(), split_comments().
}
