% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_plots.R
\name{session_plots}
\alias{session_plots}
\title{session_plots.}
\usage{
session_plots(dat, path)
}
\arguments{
\item{dat}{a iox dataframe.}

\item{path}{where to save the files}
}
\value{
plots
}
\description{
session_plots create plotz of iox objects
}
