% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_plots.R
\name{session_plots}
\alias{session_plots}
\title{session_plots.}
\usage{
session_plots(dat, inter = TRUE, path, vent_stat = "mean",
  baseline = 30, bin = 3, fsave = TRUE)
}
\arguments{
\item{dat}{a list of  dataframes as returned by \code{\link{import_session}}.}

\item{inter}{logical that determines if dialogs will be used.}

\item{path}{where to save the files.}

\item{vent_stat}{stat used to summarize bins displayed in the y axis ("median" or "mean").}

\item{baseline}{length of baseline in minutes.}

\item{bin}{length of bin in minutes.}

\item{fsave}{if TRUE save plots in pdf files in path.}
}
\value{
plots
}
\description{
session_plots creates plots from list of vent dataframes.
}
