% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_vent.R
\name{summarize_vent}
\alias{summarize_vent}
\title{summarize _vent}
\usage{
summarize_vent(dat, inter = TRUE, baseline = 30, bin = 3,
  form = "mean")
}
\arguments{
\item{dat}{a list of vent dataframes as returned by import_session.}

\item{inter}{logical.}

\item{baseline}{length of baseline in minutes.}

\item{bin}{length of bins in minutes.}

\item{form}{stat for summarizing the data.}
}
\value{
a list of dataframes:
\enumerate{
  \item \strong{dat_long}: a dataframe of the session that contains a column *measure* and bins.
  \item \strong{dat_vent}: a vent dataframe, in which summary stats for each bin have been calculated.
  \item \strong{dat_sml}: a dataframe binned and summarized in which the summary stats have been melted in the column *int_stat*.
  \item \strong{dat_fs} a dat_sm dataframe splitted by *measure*.
}
}
\description{
summarize_vent creates duration intervals, filters and summarizes a list of vent objects returned by import_vent.
}
