% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent.R
\name{vent}
\alias{vent}
\alias{vent.default}
\title{Methods of class vents}
\usage{
vent(x, ...)

\method{vent}{default}(dat)
}
\description{
These functions provide methods for collection, analyzing and visualizing a
set of vent results from a common data set
}
