% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vent.R
\name{vent}
\alias{vent}
\alias{autoplot.vent,}
\alias{new_vent,}
\alias{validate_vent,}
\alias{as.data.frame.vent}
\title{Class vent}
\usage{
vent(dat, ...)
}
\arguments{
\item{dat}{an object of class dataframe}

\item{...}{dots}
}
\value{
an object of class vent
}
\description{
Method, constructor and validator of objects of class vent
}
